(PC-34102)[PRO] refactor: Moves single test files out of /__specs__/ + lint
